Updates to ExcludeRegionPlugin:   * Added script hook to ensure that exclusion cleanup actions are performed on successful print completion, if needed   * Updated enterExcludedRegion to be NOP if already excluding   * Updated exitExcludedRegion to be NOP if not excluding   * Split tests for hooks into separate test file   * Split tests for settings into separate test file
